Add support for unified diff comments.
The first character of a line in a unified diff hunk is used to
determine the meaning of the line (added, removed or unchanged).

This proposes a new leading '#' character to indicate a comment on the
preceding line. This is a complete fabrication of my own making and may
not be standard and might break all sorts of diff tools, but I have a
plan for it -- annotating a diff with pull request comments and then
pushing these to a code review system such as git-annotate or some
centralized service via APIs.

Doesn't impact the parsing of standard unified diffs.